Improve / fix the NoIO event manager.
In previous versions of this manager, there was the possibility of thread
starvation when a GHC thread run on the same capability as the IO manager
thread. In this case, when the IO manager ran out of things to do, it would
block the entire OS / CPU thread until the next event, without checking to
see if there were other GHC threads running. The modified code now does an
explicit yield in order to catch these possible other threads.